Add channels argument, README, and directories
I think that I'm happy with the project as it is now.
